pyproject.toml
[project]
name = "my-fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.121.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
